refactor: NumericSelector 컴포넌트 타입 지정
